Refactor: Separate OCR logic into dedicated ocr.rs module
- Create src/ocr.rs with OCR-specific functions:
  * extract_text_from_image: Tesseract OCR text extraction
  * is_supported_image_format: Image format validation
- Update src/bot.rs to use crate::ocr::* for OCR operations
- Update src/main.rs to include mod ocr;
- Improve code organization and separation of concerns
- Make OCR functionality more modular and testable
